sgin system update javascript version
