+++ src/main/webapp/common/navbar.jsp
@@ -29,9 +29,9 @@
 				<li class = "nav-item"><a class = "nav-link" href =
 						"${pageContext.request.contextPath}/pages/introduction/introofindex.jsp">소개글</a>
 				</li>
-				<li class = "nav-item"><a class = "nav-link"
+				<li class = "nav-item"><a class = "nav-link" id="nav-login"
 				                          href = "${pageContext.request.contextPath}/pages/signIn.jsp">로그인</a></li>
-				<li class = "nav-item"><a class = "nav-link"
+				<li class = "nav-item"><a class = "nav-link" id="nav-signup"
 				                          href = "${pageContext.request.contextPath}/pages/signUp.jsp">회원가입</a></li>
 
 				<!-- 다크모드 토글 버튼 -->
@@ -46,3 +46,6 @@
 		</div>
 	</div>
 </nav>
+<!-- 로그인 감지해서 로그인 회원가입 바꿔치기
+<script src="${pageContext.request.contextPath}/js/common.js"></script>
+-->
