+++ src/main/webapp/pages/signUp.jsp
@@ -3,173 +3,41 @@
 <html lang="en">
 <head>
     <title>Sign Up</title>
-    <link href="../css/styles.css" rel="stylesheet"/>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"/>
     <link href="${pageContext.request.contextPath}/css/styles.css" rel="stylesheet"/>
-
-    <style>
-        body {
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            height: 100vh;
-            /* 라이트 모드 배경은 body에 직접 지정되어 있으므로, 여기서는 기본 스타일만 유지 */
-        }
-
-        .signup-container {
-            display: flex;
-            flex-direction: column;
-            justify-content: center;
-            align-items: center;
-            width: 680px;
-            padding: 40px;
-            border-radius: 12px;
-            border: 1px solid black; /* 라이트 모드 테두리 */
-            background-color: white; /* 라이트 모드 배경 (기존 yellow 대신 흰색으로 변경) */
-            color: black; /* 라이트 모드 글자색 */
-        }
-
-        /* ---------------------------------
-           다크 모드: 컨테이너 스타일 오버라이드
-           --------------------------------- */
-        body[data-mode="dark"] .signup-container {
-            background-color: #1e1e1e; /* 어두운 배경 */
-            border: 1px solid #444; /* 밝은 테두리 */
-            color: #eee; /* 밝은 글자색 */
-        }
-
-        .signup-container h2 {
-            margin-bottom: 20px;
-            color: black; /* 라이트 모드 글자색 */
-        }
-
-        /* ---------------------------------
-           다크 모드: 제목 스타일 오버라이드
-           --------------------------------- */
-        body[data-mode="dark"] .signup-container h2 {
-            color: #eee; /* 다크 모드 글자색 */
-        }
-
-        .signup-container input[type="text"],
-        .signup-container input[type="password"],
-        .signup-container input[type="email"] {
-            width: 100%;
-            padding: 12px;
-            margin-bottom: 15px;
-            border: 1px solid #ccc; /* 라이트 모드 테두리 (기존 black 대신 연한 회색으로 변경) */
-            border-radius: 6px;
-            box-sizing: border-box;
-            font-size: 16px;
-            background-color: #fff; /* 라이트 모드 배경 */
-            color: #000; /* 라이트 모드 글자색 */
-        }
-
-        /* ---------------------------------
-           다크 모드: 입력 필드 스타일 오버라이드
-           --------------------------------- */
-        body[data-mode="dark"] .signup-container input[type="text"],
-        body[data-mode="dark"] .signup-container input[type="password"],
-        body[data-mode="dark"] .signup-container input[type="email"]
-        {
-            background-color: #212121;
-            color: #eee;
-            border: 1px solid #555;
-        }
-
-        .signup-container .radio-group {
-            display: flex;
-            justify-content: center;
-            padding-left:10px;
-            width: 100%;
-            margin-bottom: 15px;
-        }
-
-        .signup-container .radio-group label {
-            margin-right: 20px;
-        }
-
-        /* 라디오 버튼 텍스트도 컨테이너 글자색을 따라가므로 별도 설정 불필요 */
-
-        .signup-container .terms-checkbox {
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            width: 100%;
-            margin-bottom: 20px;
-        }
-
-        .signup-container .terms-checkbox input {
-            margin-right: 10px;
-        }
-
-        /* ---------------------------------
-           버튼 스타일 (다크 모드에 맞춰 색상 조정)
-           --------------------------------- */
-        .signup-container button {
-            width: 100%;
-            padding: 15px;
-            border: none;
-            border-radius: 6px;
-            background-color: skyblue; /* 라이트 모드 기본 */
-            color: white;
-            font-size: 18px;
-            cursor: pointer;
-            transition: background-color 0.3s;
-        }
-
-        .signup-container button:hover {
-            background-color: blue; /* 라이트 모드 호버 */
-        }
-
-        body[data-mode="dark"] .signup-container button {
-            background-color: #007bff; /* 다크 모드 기본 */
-        }
-
-        body[data-mode="dark"] .signup-container button:hover {
-            background-color: #0056b3; /* 다크 모드 호버 */
-        }
-    </style>
+    <link href="${pageContext.request.contextPath}/css/features/sign/signStyle.css" rel="stylesheet"/>
 </head>
-<body id="pageBody" class="d-flex flex-column h-100"
-      style="background-color:#fff; color:#000;" data-mode="light">
+<body id="pageBody" class="sign-page-body" data-mode="light">
 
 <%@ include file="/common/navbar.jsp" %>
 <%@ include file="/common/sidebar.jsp" %>
 
-<!-- Sign Up Form -->
-<div class="signup-container">
+<div class="sign-container signup-container">
     <h2>회원가입</h2>
-    
-        <%-- 서버로부터 받은 에러 메시지 표시 --%>
-    <% 
+
+    <%
         String error = request.getParameter("error");
         String errorMessage = "";
         if (error != null) {
             switch (error) {
-                case "id_duplicate":
-                    errorMessage = "이미 사용 중인 아이디입니다.";
-                    break;
-                case "email_duplicate":
-                    errorMessage = "이미 등록된 이메일입니다.";
-                    break;
-                case "validation_fail":
-                    errorMessage = "입력 형식이 올바르지 않습니다.";
-                    break;
+                case "id_duplicate": errorMessage = "이미 사용 중인 아이디입니다."; break;
+                case "email_duplicate": errorMessage = "이미 등록된 이메일입니다."; break;
+                case "validation_fail": errorMessage = "입력 형식이 올바르지 않습니다."; break;
             }
         }
     %>
     <% if (!errorMessage.isEmpty()) { %>
-        <div class="error-message"><%= errorMessage %></div>
+    <div class="error-message"><%= errorMessage %></div>
     <% } %>
-    
+
     <form name="signupForm" action="${pageContext.request.contextPath}/LoginServlet" method="post">
-    	<input type="hidden" name="action" value="signup">
+        <input type="hidden" name="action" value="signup">
         ID<input type="text" name="userId" id="userId" placeholder="아이디를 입력해주세요." required>
         비밀번호<input type="password" name="userPassword" id="userPassword" placeholder="비밀번호를 입력해주세요." required>
         비밀번호 확인<input type="password" name="passwordChk" id="passwordChk" placeholder="비밀번호를 입력해주세요." required>
-        이메일<input type="email" name="email" id="email" placeholder="이메일 주소를 입력해주세요. ~ @ ~ . ~ 형식이어야합니다." required>
+        이메일<input type="email" name="email" id="email" placeholder="이메일 주소를 입력해주세요." required>
         이름<input type="text" name="name" id="name" placeholder="이름을 입력해주세요." required>
-        주소<input type="text" name="address" id="address" placeholder="주소를 입력해주세요. ~시여야합니다." required>
+        주소<input type="text" name="address" id="address" placeholder="주소를 입력해주세요." required>
 
         <div class="radio-group">
             <input type="radio" id="male" name="gender" value="male" checked>
@@ -189,11 +57,7 @@
 
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
 <script src="${pageContext.request.contextPath}/js/features/darkmode.js"></script>
-
-<script>
-    const CONTEXT_PATH = "${pageContext.request.contextPath}";
-</script>
-
-<script defer src="${pageContext.request.contextPath}/js/featuers/signUpAndValidate.js"></script>
+<script src="${pageContext.request.contextPath}/js/auth.js"></script>
+<script defer src="${pageContext.request.contextPath}/js/features/signUpAndValidate.js"></script>
 </body>
 </html>
